Implemented react Query but not working with Zustand and have fixed the Add Article Modal in this one.
