+++ package.json
@@ -7,17 +7,22 @@
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.13.1",
+    "@tanstack/react-query": "^4.29.7",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "antd": "^5.4.7",
+    "axios": "^1.4.0",
+    "formik": "^2.2.9",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-query": "^3.39.3",
     "react-router": "^6.11.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.1",
     "react-toastify": "8.0.2",
     "web-vitals": "^2.1.4",
+    "yup": "^1.1.1",
     "zustand": "^4.3.8"
   },
   "scripts": {
