+++ src/App.js
@@ -1,22 +1,26 @@
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ARITCLE_ROUTES } from "./Components/routeConstant";
 
 function App() {
+  const queryClient = new QueryClient();
   return (
-    <BrowserRouter>
-      <ToastContainer />
-      <Switch>
-        {ARITCLE_ROUTES.map((route) => (
-          <Route
-            key={route.path}
-            path={route.path}
-            exact
-            component={route.component}
-          />
-        ))}
-      </Switch>
-    </BrowserRouter>
+    <QueryClientProvider client={queryClient}>
+      <BrowserRouter>
+        <ToastContainer />
+        <Switch>
+          {ARITCLE_ROUTES.map((route) => (
+            <Route
+              key={route.path}
+              path={route.path}
+              exact
+              component={route.component}
+            />
+          ))}
+        </Switch>
+      </BrowserRouter>
+    </QueryClientProvider>
   );
 }
